pyproject.toml
[tool.poetry]
name = "streamlit-app"
version = "0.1.0"
description = ""
authors = ["Princekrampah <jsksprince@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.34.0"
llama-index = "^0.10.37"
llama-index-llms-mistralai = "^0.1.12"
python-decouple = "^3.8"
pdfkit = "^1.0.0"
streamlit-card = "^1.0.2"
streamlit-modal = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
